+++ src-tauri/src/Open_and_Save.rs
@@ -1,7 +1,7 @@
 use crate::{
     file_format::{
-        Ainb_py::Ainb_py, Asb_py::{Asb_py, ASB_SEPARATOR}, BinTextFile::{is_banc_path, replace_rotate_deg_to_rad, BymlFile, OpenedFile}, Esetb::Esetb, Evfl_cs::Evfl, Msbt::str_endian_to_roead, Pack::{PackComparer, PackFile, SarcPaths}, Rstb::Restbl, TagProduct::TagProduct, Wrapper::PythonWrapper, SMO::SmoSaveFile::SmoSaveFile
-    }, Comparer::DiffComparer, Settings::Pathlib, TotkApp::InternalFile, Zstd::{is_aamp, is_ainb, is_ainb_path, is_asb_path, is_byml, is_byml_path, is_esetb_path, is_evfl_path, is_gamedatalist, is_msbt_path, is_msyt, is_rstb_path, is_tagproduct_path, TotkFileType, TotkZstd}
+        Ainb_py::Ainb_py, Asb_py::{Asb_py, ASB_SEPARATOR}, BinTextFile::{is_banc_path, replace_rotate_deg_to_rad, BymlFile, OpenedFile}, Esetb::Esetb, Evfl_cs::Evfl, Msbt::str_endian_to_roead, Pack::{PackComparer, PackFile, SarcPaths}, Rstb::Restbl, TagProduct::TagProduct, Wrapper::PythonWrapper, Xlink::Xlink_rs, SMO::SmoSaveFile::SmoSaveFile
+    }, Comparer::DiffComparer, Settings::Pathlib, TotkApp::InternalFile, Zstd::{is_aamp, is_ainb, is_ainb_path, is_asb_path, is_byml, is_byml_path, is_esetb_path, is_evfl_path, is_gamedatalist, is_msbt_path, is_msyt, is_rstb_path, is_tagproduct_path, is_xlink_path, TotkFileType, TotkZstd}
 };
 use msbt_bindings_rs::MsbtCpp::MsbtCpp;
 use rfd::{FileDialog, MessageDialog};
@@ -41,162 +41,6 @@ pub fn open_sarc<P: AsRef<Path>>(
 
     None
 }
-pub fn open_esetb<P:AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
-    let mut opened_file = OpenedFile::default();
-    let path_ref = path.as_ref();
-    let mut data = SendData::default();
-    print!("Is {:?} a esetb? ", &path_ref);
-    if is_esetb_path(&path) {
-        opened_file.esetb = Esetb::from_file(path_ref, zstd.clone()).ok();
-        if let Some(esetb) = &opened_file.esetb {
-            println!(" yes!");
-            data.tab = "YAML".to_string();
-            opened_file.path = Pathlib::new(path_ref);
-            opened_file.endian = esetb.byml.endian;
-            opened_file.file_type = TotkFileType::Esetb;
-            data.status_text = format!("Opened {}", path_ref.display());
-            data.path = Pathlib::new(path_ref);
-            data.text = esetb.to_string();
-            data.get_file_label(TotkFileType::Esetb, esetb.byml.endian);
-            return Some((opened_file, data));
-        }
-    }
-    println!("no");
-
-    None
-}
-pub fn open_restbl<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
-    let mut opened_file = OpenedFile::default();
-    let path_ref = path.as_ref();
-    let mut data = SendData::default();
-    print!("Is {} a restbl? ", &path_ref.display());
-    let pathlib_var = Pathlib::new(path_ref);
-    if pathlib_var
-        .name
-        .to_lowercase()
-        .starts_with("resourcesizetable.product")
-    {
-        println!(" yes!");
-        opened_file.restbl = Restbl::from_path(path_ref, zstd.clone());
-        if let Some(_restbl) = &mut opened_file.restbl {
-            data.tab = "RSTB".to_string();
-            opened_file.path = pathlib_var.clone();
-            opened_file.endian = Some(roead::Endian::Little);
-            opened_file.file_type = TotkFileType::Restbl;
-            data.status_text = format!("Opened {}", &pathlib_var.full_path);
-            data.path = pathlib_var;
-            // data.text = restbl.to_text();
-            data.get_file_label(TotkFileType::Restbl, Some(roead::Endian::Little));
-            return Some((opened_file, data));
-        }
-    }
-    println!(" no");
-    None
-}
-
-pub fn open_tag<P:AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
-    let mut opened_file = OpenedFile::default();
-    let mut data = SendData::default();
-    let path_ref = path.as_ref();
-    let pathlib_var = Pathlib::new(path_ref);
-    print!("Is {} a tag? ", &pathlib_var.full_path);
-    if is_tagproduct_path(path_ref)
-    {
-        opened_file.tag = TagProduct::new(path_ref, zstd.clone());
-        if let Some(tag) = &mut opened_file.tag {
-            println!(" yes!");
-            opened_file.path = pathlib_var.clone();
-            opened_file.endian = Some(roead::Endian::Little);
-            opened_file.file_type = TotkFileType::TagProduct;
-            data.status_text = format!("Opened {}", &pathlib_var.full_path);
-            data.path = pathlib_var;
-            data.text = tag.to_text();
-            data.lang = "json".to_string();
-            data.get_file_label(TotkFileType::TagProduct, Some(roead::Endian::Little));
-            return Some((opened_file, data));
-        }
-    }
-    println!(" no");
-    None
-}
-
-pub fn open_asb<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
-    let mut opened_file = OpenedFile::default();
-    let path_ref = path.as_ref();
-    let mut data = SendData::default();
-    print!("Is {} a asb? ", &path_ref.display());
-    if let Ok(asb) = Asb_py::from_binary_file(path_ref, zstd.clone()) {
-        match asb.binary_to_text()  {
-            Ok(text) => {
-                println!(" yes!");
-                opened_file.path = Pathlib::new(path_ref);
-                opened_file.file_type = TotkFileType::ASB;
-                data.status_text = format!("Opened: {}", &opened_file.path.full_path);
-                data.path = Pathlib::new(path_ref);
-                data.text = text;
-                data.get_file_label(TotkFileType::ASB, Some(roead::Endian::Little));
-                return Some((opened_file, data));
-            }
-            Err(e) => {
-                println!(" yes but failed to open: {}", e);
-            }
-        }
-    }
-    println!(" no");
-    None
-}
-
-pub fn open_ainb<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
-    let mut opened_file = OpenedFile::default();
-    let mut data = SendData::default();
-    let path_ref = path.as_ref();
-    print!("Is {} a ainb? ", &path_ref.display());
-    match Ainb_py::new().binary_file_to_text(path_ref) {
-        Ok(text) => {
-            println!(" yes!");
-            opened_file.path = Pathlib::new(path_ref);
-            opened_file.file_type = TotkFileType::AINB;
-            data.status_text = format!("Opened: {}", &opened_file.path.full_path);
-            data.path = Pathlib::new(path_ref);
-            data.text = text;
-            data.get_file_label(TotkFileType::AINB, None);
-            return Some((opened_file, data));
-        }
-        Err(e) => {
-            println!(" no: {}", e);
-        }
-    }
-    None
-}
-pub fn open_byml<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
-    let mut opened_file = OpenedFile::default();
-    let mut data = SendData::default();
-    let path_ref = path.as_ref();
-    let pathlib_var = Pathlib::new(path_ref);
-    print!("Is {} a byml? ", &pathlib_var.full_path);
-    opened_file.byml = BymlFile::new(path_ref, zstd.clone());
-    // if opened_file.byml.is_some() {
-    if let Some(b) = &opened_file.byml {
-        // let b = opened_file.byml.as_ref().unwrap();
-        let gamedatalist = if is_gamedatalist(path_ref) {
-            "(GameDataList) "
-        } else {
-            ""
-        };
-        println!("yes {}!",  gamedatalist);
-        opened_file.path = pathlib_var.clone();
-        opened_file.endian = b.endian;
-        opened_file.file_type = b.file_data.file_type.clone();
-        data.status_text = format!("Opened {}", &pathlib_var.full_path);
-        data.path = pathlib_var;
-        // data.text = Byml::to_text(&b.pio);
-        data.text = b.to_string();
-        data.get_file_label(b.file_data.file_type, b.endian);
-        return Some((opened_file, data));
-    }
-    println!(" no");
-    None
-}
 
 pub fn open_msbt<P:AsRef<Path>>(path: P) -> Option<(OpenedFile<'static>, SendData)> {
     let file_name = path.as_ref().to_string_lossy().to_string().replace("\\", "/");
@@ -221,29 +65,6 @@ pub fn open_msbt<P:AsRef<Path>>(path: P) -> Option<(OpenedFile<'static>, SendDat
     None
 }
 
-pub fn open_smo_save_file<P:AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile<'static>, SendData)> {
-    let file_name = path.as_ref().to_string_lossy().to_string().replace("\\", "/");
-    let mut opened_file = OpenedFile::default();
-    let mut data = SendData::default();
-    let pathlib_var = Pathlib::new(&file_name);
-    print!("Is {} a smo save file?", &file_name);
-    if let Ok(smo_file) = &mut SmoSaveFile::from_file(&file_name, zstd.clone()) {
-        if let Ok(text) = smo_file.to_string() {
-            println!(" yes!");
-            opened_file.path = pathlib_var.clone();
-            opened_file.endian = Some(smo_file.endian);
-            opened_file.file_type = TotkFileType::SmoSaveFile;
-            data.status_text = format!("Opened {}", &pathlib_var.full_path);
-            data.path = pathlib_var;
-            data.text = text;
-            data.get_file_label(opened_file.file_type, Some(smo_file.endian));
-            return Some((opened_file, data));
-        }
-        // let m = opened_file.msyt.as_ref().unwrap();
-    }
-    println!(" no");
-    None
-}
 
 pub fn open_text<P: AsRef<Path>>(path: P) -> Option<(OpenedFile<'static>, SendData)> {
     let mut opened_file = OpenedFile::default();
@@ -446,6 +267,15 @@ pub fn get_binary_by_filetype(
     let is_zs = file_path.to_lowercase().ends_with(".zs");
     let is_bcett = file_path.to_lowercase().ends_with(".bcett.byml.zs");
     match file_type {
+        TotkFileType::Xlink => {
+            let xlink = Xlink_rs::new(zstd.clone()).ok()?;
+            if let Ok(new_data) = xlink.yaml_to_binary(text) {
+                rawdata = new_data;
+                if is_zs {
+                    rawdata = zstd.compress_zs(&rawdata).ok()?;
+                }
+            }
+        }
         TotkFileType::Evfl => {
           let evfl = Evfl::new(zstd.clone());
           if let Ok(new_data) = evfl.string_to_binary(text) {
@@ -797,25 +627,25 @@ pub fn file_from_disk_to_senddata_determine_ext<P: AsRef<Path>>(path: P, zstd: A
     }
     let path = Pathlib::new(&p);
     if is_tagproduct_path(&file_name) {
-        return open_tag(&file_name, zstd.clone());
+        return TagProduct::open_tag(&file_name, zstd.clone());
     }
     if is_esetb_path(&file_name) {
-        return open_esetb(&file_name, zstd.clone());
+        return Esetb::open_esetb(&file_name, zstd.clone());
+    }
+    if is_xlink_path(&p) {
+        return Xlink_rs::open_xlink(&file_name, zstd.clone());
     }
     if is_ainb_path(&p) {
-        return open_ainb(&file_name, zstd.clone());
+        return Ainb_py::open_ainb(&file_name, zstd.clone());
     }
     if is_rstb_path(&p) {
-        return open_restbl(&file_name, zstd.clone());
+        return Restbl::open_restbl(&file_name, zstd.clone());
     }
     if is_asb_path(&p) {
-        return open_asb(&file_name, zstd.clone());
-    }
-    if is_asb_path(&p) {
-        return open_asb(&file_name, zstd.clone());
+        return Asb_py::open_asb(&file_name, zstd.clone());
     }
     if is_byml_path(&p) {
-        return open_byml(&file_name, zstd.clone());
+        return BymlFile::open_byml(&file_name, zstd.clone());
     }
     if is_msbt_path(&p) {
         return open_msbt(&file_name);
@@ -826,20 +656,22 @@ pub fn file_from_disk_to_senddata_determine_ext<P: AsRef<Path>>(path: P, zstd: A
     
     None // if fails, proceed to regular file open
 }
+
 pub fn file_from_disk_to_senddata<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
     let file_name = path.as_ref();//.to_string_lossy().to_string().replace("\\", "/");
     // let res = open_tag(&file_name, zstd.clone())
     let res = file_from_disk_to_senddata_determine_ext(&file_name, zstd.clone())
-                .or_else(|| open_tag(&file_name, zstd.clone()))
-                .or_else(|| open_esetb(&file_name, zstd.clone()))
-                .or_else(|| open_restbl(&file_name, zstd.clone()))
-                .or_else(|| open_asb(&file_name, zstd.clone()))
-                .or_else(|| open_ainb(&file_name, zstd.clone()))
-                .or_else(|| open_byml(&file_name, zstd.clone()))
+                .or_else(|| TagProduct::open_tag(&file_name, zstd.clone()))
+                .or_else(|| Esetb::open_esetb(&file_name, zstd.clone()))
+                .or_else(|| Xlink_rs::open_xlink(&file_name, zstd.clone()))
+                .or_else(|| Restbl::open_restbl(&file_name, zstd.clone()))
+                .or_else(|| Asb_py::open_asb(&file_name, zstd.clone()))
+                .or_else(|| Ainb_py::open_ainb(&file_name, zstd.clone()))
+                .or_else(|| BymlFile::open_byml(&file_name, zstd.clone()))
                 .or_else(|| open_msbt(&file_name))
                 .or_else(|| open_aamp(&file_name))
                 .or_else(|| Evfl::open_file(&file_name, zstd.clone()))
-                .or_else(|| open_smo_save_file(&file_name, zstd.clone()))
+                .or_else(|| SmoSaveFile::open_smo_save_file(&file_name, zstd.clone()))
                 .or_else(|| open_text(&file_name))
                 .map(|(opened_file, data)| {
                     // self.opened_file = opened_file;