+++ src-tauri/src/file_format/Ainb_py.rs
@@ -1,10 +1,12 @@
 #![allow(non_snake_case,non_camel_case_types)]
 use std::{
-    io::{self, Read, Write}, os::windows::process::CommandExt, path::Path, process::{Command, Stdio}
+    io::{self, Read, Write}, os::windows::process::CommandExt, path::Path, process::{Command, Stdio}, sync::Arc
 };
 
 
-use crate::{Settings::NO_WINDOW_FLAG, Zstd::is_ainb};
+use crate::{Open_and_Save::SendData, Settings::{Pathlib, NO_WINDOW_FLAG}, Zstd::{is_ainb, TotkFileType, TotkZstd}};
+
+use super::BinTextFile::OpenedFile;
 
 pub struct Ainb_py {
     pub python_exe: String,
@@ -27,6 +29,30 @@ impl Ainb_py {
     pub fn new() -> Self {
         Self::default()
     }
+
+    pub fn open_ainb<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
+        let mut opened_file = OpenedFile::default();
+        let mut data = SendData::default();
+        let path_ref = path.as_ref();
+        print!("Is {} a ainb? ", &path_ref.display());
+        match Ainb_py::new().binary_file_to_text(path_ref) {
+            Ok(text) => {
+                println!(" yes!");
+                opened_file.path = Pathlib::new(path_ref);
+                opened_file.file_type = TotkFileType::AINB;
+                data.status_text = format!("Opened: {}", &opened_file.path.full_path);
+                data.path = Pathlib::new(path_ref);
+                data.text = text;
+                data.get_file_label(TotkFileType::AINB, None);
+                return Some((opened_file, data));
+            }
+            Err(e) => {
+                println!(" no: {}", e);
+            }
+        }
+        None
+    }
+
     pub fn binary_file_to_text<P:AsRef<Path>>(&self, file_path: P) -> io::Result<String> {
         // env::set_var("PATH", self.newpath.clone());
         let mut f_handle = std::fs::File::open(file_path)?; // Open the file