+++ src-tauri/src/file_format/Asb_py.rs
@@ -1,6 +1,8 @@
 #![allow(non_snake_case, non_camel_case_types)]
+use crate::file_format::BinTextFile::OpenedFile;
+use crate::Open_and_Save::SendData;
 use crate::Settings::{Pathlib, NO_WINDOW_FLAG};
-use crate::Zstd::{is_asb, is_baev, TotkZstd};
+use crate::Zstd::{is_asb, is_baev, TotkFileType, TotkZstd};
 use rfd::{FileDialog, MessageDialog};
 use std::fs;
 use std::path::{Path, PathBuf};
@@ -307,4 +309,32 @@ impl<'a> Asb_py<'a> {
         println!("Test response from winpython: {}", text);
         Ok(())
     }
+
+    pub fn open_asb<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
+        let mut opened_file = OpenedFile::default();
+        let path_ref = path.as_ref();
+        let mut data = SendData::default();
+        print!("Is {} a asb? ", &path_ref.display());
+        if let Ok(asb) = Asb_py::from_binary_file(path_ref, zstd.clone()) {
+            match asb.binary_to_text()  {
+                Ok(text) => {
+                    println!(" yes!");
+                    opened_file.path = Pathlib::new(path_ref);
+                    opened_file.file_type = TotkFileType::ASB;
+                    data.status_text = format!("Opened: {}", &opened_file.path.full_path);
+                    data.path = Pathlib::new(path_ref);
+                    data.text = text;
+                    data.get_file_label(TotkFileType::ASB, Some(roead::Endian::Little));
+                    return Some((opened_file, data));
+                }
+                Err(e) => {
+                    println!(" yes but failed to open: {}", e);
+                }
+            }
+        }
+        println!(" no");
+        None
+    }
+    
+
 }