+++ src-tauri/src/file_format/BinTextFile.rs
@@ -1,5 +1,6 @@
 #![allow(non_snake_case,non_camel_case_types)]
 use crate::file_format::TagProduct::TagProduct;
+use crate::Open_and_Save::SendData;
 use crate::Settings::Pathlib;
 use crate::Zstd::{is_byml, is_gamedatalist, TotkFileType, TotkZstd};
 use msbt_bindings_rs::MsbtCpp::MsbtCpp;
@@ -239,7 +240,35 @@ impl<'a> BymlFile<'_> {
         // process_inline_content(Byml::to_text(&self.pio), self.zstd.totk_config.yaml_max_inl)
         text
     }
-
+    pub fn open_byml<P: AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
+        let mut opened_file = OpenedFile::default();
+        let mut data = SendData::default();
+        let path_ref = path.as_ref();
+        let pathlib_var = Pathlib::new(path_ref);
+        print!("Is {} a byml? ", &pathlib_var.full_path);
+        opened_file.byml = BymlFile::new(path_ref, zstd.clone());
+        // if opened_file.byml.is_some() {
+        if let Some(b) = &opened_file.byml {
+            // let b = opened_file.byml.as_ref().unwrap();
+            let gamedatalist = if is_gamedatalist(path_ref) {
+                "(GameDataList) "
+            } else {
+                ""
+            };
+            println!("yes {}!",  gamedatalist);
+            opened_file.path = pathlib_var.clone();
+            opened_file.endian = b.endian;
+            opened_file.file_type = b.file_data.file_type.clone();
+            data.status_text = format!("Opened {}", &pathlib_var.full_path);
+            data.path = pathlib_var;
+            // data.text = Byml::to_text(&b.pio);
+            data.text = b.to_string();
+            data.get_file_label(b.file_data.file_type, b.endian);
+            return Some((opened_file, data));
+        }
+        println!(" no");
+        None
+    }
 
 }
 
@@ -473,45 +502,7 @@ impl<'a> OpenedFile<'_> {
         }
     }
 
-    pub fn open(&mut self, file_path: &str, zstd: Arc<TotkZstd>) -> String {
-        let res = String::new();
-        let path = Pathlib::new(file_path.to_string());
-        if self.open_tag(&path, zstd.clone()) {
-            if let Some(tag) = &self.tag {
-                return tag.text.to_string();
-            }
-        }
-        res
-    }
-
-    pub fn open_tag(&mut self, path: &Pathlib, zstd: Arc<TotkZstd>) -> bool {
-        if path.name.to_lowercase().starts_with("tag.product") {
-            let tag = TagProduct::new(path.full_path.clone(), zstd.clone());
-            match tag {
-                Some(mut tag) => {
-                    match tag.parse() {
-                        Ok(_) => {
-                            println!("Tag parsed!");
-                        }
-                        Err(err) => {
-                            eprintln!("Error parsing tag! {:?}", err);
-                            return false;
-                        }
-                    }
-                    self.reset();
-                    //self.tag = Some(tag);
-                    self.file_type = TotkFileType::TagProduct;
-                    self.path = path.clone();
-                    self.endian = Some(roead::Endian::Little);
-                    return true;
-                }
-                None => {
-                    return false;
-                }
-            }
-        }
-        false
-    }
+    
 }
 
 #[allow(dead_code)]