+++ src-tauri/src/file_format/TagProduct.rs
@@ -1,6 +1,8 @@
 #![allow(non_snake_case, non_camel_case_types)]
 use crate::file_format::BinTextFile::{bytes_to_file, BymlFile};
-use crate::Zstd::TotkZstd;
+use crate::Open_and_Save::SendData;
+use crate::Settings::Pathlib;
+use crate::Zstd::{is_tagproduct_path, TotkFileType, TotkZstd};
 //use byteordered::Endianness;
 //use indexmap::IndexMap;
 use bitvec::prelude::*;
@@ -14,6 +16,8 @@ use std::panic::AssertUnwindSafe;
 use std::path::Path;
 use std::sync::Arc;
 use std::{io, panic};
+
+use super::BinTextFile::OpenedFile;
 
 #[derive(Serialize, Deserialize)]
 struct TagJsonData {
@@ -286,6 +290,32 @@ impl<'a> TagProduct<'a> {
             //self.to_text();
         }
         Ok(())
+    }
+
+    pub fn open_tag<P:AsRef<Path>>(path: P, zstd: Arc<TotkZstd>) -> Option<(OpenedFile, SendData)> {
+        let mut opened_file = OpenedFile::default();
+        let mut data = SendData::default();
+        let path_ref = path.as_ref();
+        let pathlib_var = Pathlib::new(path_ref);
+        print!("Is {} a tag? ", &pathlib_var.full_path);
+        if is_tagproduct_path(path_ref)
+        {
+            opened_file.tag = TagProduct::new(path_ref, zstd.clone());
+            if let Some(tag) = &mut opened_file.tag {
+                println!(" yes!");
+                opened_file.path = pathlib_var.clone();
+                opened_file.endian = Some(roead::Endian::Little);
+                opened_file.file_type = TotkFileType::TagProduct;
+                data.status_text = format!("Opened {}", &pathlib_var.full_path);
+                data.path = pathlib_var;
+                data.text = tag.to_text();
+                data.lang = "json".to_string();
+                data.get_file_label(TotkFileType::TagProduct, Some(roead::Endian::Little));
+                return Some((opened_file, data));
+            }
+        }
+        println!(" no");
+        None
     }
 }