+++ src-tauri/src/file_format/mod.rs
@@ -13,3 +13,4 @@ pub mod Esetb;
 pub mod Wrapper;
 pub mod SMO;
 pub mod Evfl_cs;
+pub mod Xlink;