+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - faster file opening
 - fixed tag.product parsing issues
+- Xlink editing support